- Refactoring guideline code
